Add a GetMatchesDynamic method for Target
This extends the idea of the NamespaceSelector to other types.

The FakePolicy tests demonstrate that Target can be used in a cluster-
scoped manner; it would be easy to restrict it to a specific namespace.
Using NamespaceSelector and Target to limit the policy to specific
resources in specific namespaces is possible, but might require a loop
over all matched namespaces, with a namespaced call in each.

With the new fields in the FakePolicy, one of the format verification
test cases was unnecessary.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>