Adds a non-namespaced GetMatches method to Target
The new method does not use a dynamic client, which allows it to utilize
the read cache that is configured by default on the controller-runtime
clients. This can improve the speed of the controller, while also
reducing load on the cluster's API server.

The FakePolicy controller demonstrates (and tests) how to implement the
new ResourceList interface, necessary because the `client.ObjectList`
interface does not provide any access to the list's items.

This is a partial implementation, which does not allow restricting the
matches by namespace (that is coming next).

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>