Add namespace field to Target type
GetMatches and GetMatchesDynamic are now able to be namespaced by the
Target. Previously, only GetMatchesDynamic could be namespaced, and it
had to be done by the user when creating the ResourceInterface. This
seems to be a more ergonomic solution.

The tests for the namespaced Targets are mostly duplicates of the tests
for the non-namespaced Targets; some refactoring allowed them to be
re-used without copying them, with only some small adjustments.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>