Add test toolkit and improve tests
The test toolkit and courtesies are things that other policies could use
in their tests. The remaining functions in the utils.go file are more
specifically just for fakepolicy.

In preparation for more tests, the existing tests have been moved into a
separate package/suite. In particular, since some of them require very
specific details in the cluster (exact namespaces, configmaps, and no
extras), this lets them be run in a more isolated way.

Signed-off-by: Justin Kulikauskas <44813129+JustinKuli@users.noreply.github.com>