Prune more kubebuilder scaffolding
The RBAC resources for PolicyCore were already incomplete, and are not
used, so they have now been removed from the config/ folder.

The boilerplate.go.txt was looking very lonely in the hack/ folder,
which was taking up valuable space on the GitHub landing page, so it has
been moved into the config/ folder.

Signed-off-by: Justin Kulikauskas <44813129+JustinKuli@users.noreply.github.com>