Publish coverage and docs to GH Pages
Previously, the coverage reports were pushed to the wiki, but since the
wiki is not meant for serving arbitrary HTML, viewing them properly
required going through `raw.githack.com`.

Also, the main `pkg.go.dev` site does not handle this repository well,
possibly because it is a fork (also complicated by the archiving of the
original) or because downloads of the package are very rare. Based on a
suggestion in github (dot) com/golang/go (slash) issues/2381, an updated
action now downloads the contents of a local documentation server with
`wget`, and modifies some local links with `sed` to get something which
can be statically served. This mostly works, the main downside is that
JS files seem to be missing.

This adds on to the existing test step in the action, because that step
is necessary for the coverage information. If additional resources are
to be served on the GH Pages site, it might be a good idea to move to
a separate workflow, even though it would require re-running the tests.

Signed-off-by: Justin Kulikauskas <44813129+JustinKuli@users.noreply.github.com>